feat(gui): migrate QuestionDialog to Qt Designer (.ui)
